feat: Separate scramblers of specific types to each class, inherited from ABstractScrambler
